Add changes for update and delete in case non-existing user (#50)
* Add test logic for update non-existing user

* Add test logic for update non-existing user

* Add test logic for update non-existing user

* Fix step comments

Add test logic for delete non-existing user

* Add test logic for update and delete non-existing user

+++ user-profile-app/src/main/java/com/iprody/userprofileservice/controllers/UserController.java
@@ -36,16 +36,18 @@ public class UserController implements V1UserApi {
 
     @Override
     public ResponseEntity<UserDto> updateUser(@Valid @RequestBody UserDto userDto) {
-        userService.findUserById(userDto.getId())
-                .orElseThrow(() -> new ResourceNotFoundException("User", userDto.getId()));
+        if (!userService.isExistUser(userDto.getId())) {
+            throw new ResourceNotFoundException("User", userDto.getId());
+        }
         userService.updateUser(userDto);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @Override
     public ResponseEntity<Void> deleteUserById(@PathVariable("id") Long userId) {
-        userService.findUserById(userId)
-                .orElseThrow(() -> new ResourceNotFoundException("User", userId));
+        if (!userService.isExistUser(userId)) {
+            throw new ResourceNotFoundException("User", userId);
+        }
         userService.deleteUserById(userId);
         return ResponseEntity.noContent().build();
     }
